backend/requirements.txt
fastapi>=0.118.0
uvicorn[standard]==0.37.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.10
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic>=2.2.0
email-validator>=2.3.0
alembic==1.16.5
python-jose==3.5.0
redis==5.0.3
pytest==8.4.2
httpx==0.28.1
